Prep for release 2.1.0:
Minimize icon is now single click to open like most other apps
Attempts to fix CurtLog/load assembly properly.
Remove fog option entirely; it does not provide any advantage when disabled (as the render distance is the same), but may give the appearance of doing so. As such, it is not fit for this application.
Remove dvar unlock option; same reason as above.
Some ultra hacky workarounds to get the main form to minimize and be invisible, but for functionality to remain (for example, this fixes a bug where if you minimize before launching the game, it seems to fail to detect and change FoV).
Further hacky fixes to make the app single-instance, but you can use -multi to run a second instance.
This is incredibly disgusting, but with WinForms, what isn't?
I don't know if a better way is possible. This is a long commit message.
Launch game via Steam if Steam is detected in the path.
CoD Picture box uses slightly different code so the game doesn't have to be fully loaded before it knows if it's VCoD or UO.
Remove empty line in HOTFIX